+++ src/components/ExtendedMultiselectInput.vue
@@ -42,7 +42,7 @@
       :style="multipleBlocksMargin"
       :disabled="disabled"
       :loading="loading"
-      :icon-filter="iconFilter"
+      :icon-filter="loaderIconFilter"
       :show-deselect-icon-loader="showDeselectIconLoader"
       :toggle-multiple-blocks-limit="toggleMultipleBlocksLimit"
       :empty-objects-placeholder="emptyObjectsPlaceholder"
@@ -91,7 +91,7 @@
     <extended-multiselect-multiple
       :disabled="disabled"
       :loading="loading"
-      :icon-filter="iconFilter"
+      :icon-filter="loaderIconFilter"
       :placeholder-block-shown="!selectedOptions.length"
       :show-deselect-icon-loader="showDeselectIconLoader"
       :toggle-multiple-blocks-limit="toggleMultipleBlocksLimit"
@@ -316,10 +316,10 @@ const props = defineProps({
   },
 
   /**
-   * Defines a svg-filter for icons
-   * @property {string} iconFilter
+   * Defines a svg-filter for loader icons
+   * @property {string} loaderIconFilter
    */
-  iconFilter: {
+  loaderIconFilter: {
     type: String,
     required: true,
   },